privatize zenoh::net, move benches around to let them build
