+++ commons/zenoh-protocol/Cargo.toml
@@ -44,3 +44,25 @@ uhlc = "0.5.0"
 [dev-dependencies]
 rand = "0.8.3"
 uuid = { version = "1.1.2", features = ["v4"] }
+criterion = "0.3.4"
+
+
+[[bench]]
+name = "codec_bench"
+harness = false
+
+[[bench]]
+name = "rwbuf_bench"
+harness = false
+
+[[bench]]
+name = "data_creation"
+harness = false
+
+[[bench]]
+name = "frame_codec"
+harness = false
+
+[[bench]]
+name = "tables_bench"
+harness = false
