+++ commons/zenoh-protocol/benches/rwbuf_bench.rs
@@ -17,8 +17,8 @@ extern crate rand;
 
 use criterion::{black_box, Criterion};
 
-use zenoh::net::protocol::core::ZInt;
-use zenoh::net::protocol::io::WBuf;
+use zenoh_protocol::core::ZInt;
+use zenoh_protocol::io::WBuf;
 use zenoh_protocol::io::WBufCodec;
 
 fn bench_foo((v, buf): (ZInt, &mut WBuf)) {
