+++ examples/examples/z_pub_shm.rs
@@ -14,7 +14,7 @@
 use async_std::task::sleep;
 use clap::{App, Arg};
 use std::time::Duration;
-use zenoh::buf::SharedMemoryManager;
+use zenoh::buffers::SharedMemoryManager;
 use zenoh::config::Config;
 use zenoh::prelude::r#async::AsyncResolve;
