+++ zenoh/Cargo.toml
@@ -104,31 +104,8 @@ uhlc = "0.5.0"
 uuid = { version = "1.1.2", features = ["v4"] }
 vec_map = "0.8.2"
 
-[dev-dependencies]
-criterion = "0.3.4"
-
 [build-dependencies]
 rustc_version = "0.4.0"
 
 [lib]
 name = "zenoh"
-
-[[bench]]
-name = "codec_bench"
-harness = false
-
-[[bench]]
-name = "rwbuf_bench"
-harness = false
-
-[[bench]]
-name = "data_creation"
-harness = false
-
-[[bench]]
-name = "frame_codec"
-harness = false
-
-[[bench]]
-name = "tables_bench"
-harness = false
