+++ zenoh/src/net/mod.rs
@@ -11,18 +11,20 @@
 // Contributors:
 //   ZettaScale Zenoh Team, <zenoh@zettascale.tech>
 //
+
 #[doc(hidden)]
-pub use zenoh_link as link;
-#[doc(hidden)]
-pub mod protocol {
+pub(crate) mod protocol {
     #[deprecated = "This module is now a separate crate. Use the crate directly for shorter compile-times"]
     pub use zenoh_protocol::{io, proto};
     #[deprecated = "This module is now a separate crate. Use the crate directly for shorter compile-times"]
     pub use zenoh_protocol_core as core;
 }
 #[doc(hidden)]
-pub mod routing;
+pub(crate) mod routing;
 #[doc(hidden)]
 pub mod runtime;
 #[doc(hidden)]
-pub use zenoh_transport as transport;
+pub(crate) use zenoh_transport as transport;
+
+#[cfg(test)]
+pub(crate) mod tests;
